\name{DJIRealizedMeasures}
\alias{DJIRealizedMeasures}
\docType{data}
\title{
Dow Jones Realized Measures
}
\description{
Realized measures
}

\format{
A large \code{xts} object containing trading data for the Dow Jones Industial index from 2001 to september 2018
}
\details{
See the website of the data set for details.
}
\source{
https://realized.oxford-man.ox.ac.uk/data
}
\references{
Heber, Gerd, Asger Lunde, Neil Shephard and Kevin Sheppard (2009) "Oxford-Man Institute's realized library", 
Oxford-Man Institute, University of Oxford. 
Library version: 0.3
\url{https://realized.oxford-man.ox.ac.uk/data}
}
\keyword{datasets}
