\name{HARModel-package}
\alias{HARModel-package}
\alias{HARModel}
\docType{package}
\title{
\packageTitle{HARModel}
}
\description{
\packageDescription{HARModel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HARModel}
\packageIndices{HARModel}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HARModel}

Maintainer: \packageMaintainer{HARModel}
}
\references{
@article{Corsi2009,
author = {Corsi, Fulvio},
doi = {10.1093/jjfinec/nbp001},
issn = {14798409},
journal = {Journal of Financial Econometrics},
title = {{A simple approximate long-memory model of realized volatility}},
volume = {7},
year = {2009}
}
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
