\name{HARModel-package}
\alias{HARModel-package}
\alias{HARModel}
\docType{package}
\title{
\packageTitle{HARModel}
}
\description{
\packageDescription{HARModel}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{HARModel}
\packageIndices{HARModel}
}
\author{
\packageAuthor{HARModel}
Maintainer: \packageMaintainer{HARModel}
}
\references{
Corsi, F. 2009, A Simple Approximate Long-Memory Model
of Realized Volatility, \emph{Journal of Financial Econometrics}, 174--196 .\cr
}
\keyword{ Heterogeneous Autoregressive model }
